ORM/Orm => Add comment for __get() ORM/Query/Counter => Unification of Counters of MySQL and PGSQL ORM/Query/Finder => Unification of Finders of MySQL and PGSQL ORM/Query/QueryBuilder => Move class QueryBuilder Template/Parser => Optimisation of preg_match in parse() Template/Extensions/Extensions/Escape/Html/EscapeHtmlParse => Remove increment and decrement level esc_html, and remove flag U in end of regex Template/Parsing/ParseWithExtend => Remove in attributes level and count, keys esc_html rm -rf ORM/QueryBuilder rm -rf ORM/MySQL/Counter rm -rf ORM/MySQL/Finder rm -rf ORM/PGSQL/Counter rm -rf ORM/PGSQL/Finder
