+++ src/Template/Parsing/ParseWithExtend.php
@@ -18,8 +18,8 @@ class ParseWithExtend extends ParseArray{
         'tpl' => '',
         'parts_in_escape' => [],
         'display' => '',
-        'level' => [ 'condition' => 0, 'esc_tpl' => 0, 'esc_html' => 0, 'for' => 0, 'foreach' => 0, 'index' => 0, 'itvl' => 0, 'part' => 0 ],
-        'count' => [ 'general' => 0, 'condition' => 0, 'esc_tpl' => 0, 'esc_html' => 0, 'for' => 0, 'foreach' => 0, 'index' => 0, 'itvl' => 0, 'part' => 0 ],
+        'level' => [ 'condition' => 0, 'esc_tpl' => 0, 'for' => 0, 'foreach' => 0, 'index' => 0, 'itvl' => 0, 'part' => 0 ],
+        'count' => [ 'general' => 0, 'condition' => 0, 'esc_tpl' => 0, 'for' => 0, 'foreach' => 0, 'index' => 0, 'itvl' => 0, 'part' => 0 ],
         'condition' => [],
         'foreach' => [],
         'function' => [],
